separate demo-specific stuff from general client stuff
