+++ html/index.html
@@ -202,5 +202,6 @@ To stop the demo, press the Stop button.
 </div>
 
 <script type="module" src="./app.js"></script>
+<script type="module" src="./demo.js"></script>
 <script type="module" src="./lyrics.js"></script>
 </html>
